Always fold child pages and refetch on expand
Do not store the folded state off a page anymore but always fetch
child pages via ajax on click on expand child page icon.

This drastically reduces load times on large page trees with the
downside of adding more clicks for the user. This is a tradeoff
worth the perfomance gains and reduces complextiy.

The page tree is a relict of old times that will be removed in the near
future anyway.